Ensure package dir matches package declaration. Small tidies.
