--- kafka-connect-influxdb2/src/main/scala/com/datamountaineer/streamreactor/connect/influx/ValidateStringParameterFn.scala
@@ -1,22 +1,0 @@
-/*
- * Copyright 2017 Datamountaineer.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.datamountaineer.streamreactor.connect.influx2
-
-object ValidateStringParameterFn {
-  def apply(value: String, parameterName: String): Unit =
-    require(value != null && value.trim.length > 0, s"Invalid $parameterName. Expecting non null non empty string")
-}
