+++ kafka-connect-influxdb2/src/main/scala/com/datamountaineer/streamreactor/connect/influx2/writers/InfluxBatchPointsBuilder.scala
