+++ kafka-connect-influxdb2/src/test/scala/com/datamountaineer/streamreactor/connect/influx2/ValuesExtractorJsonTest.scala
